requirements.txt
numpy
Pillow
matplotlib
PyWavelets
joblib
scipy
numba
scikit-image
